dseed: graphical user interface to download claims from channels
The program is called `dseed.py`; it is made in Python
with the `tkinter` (Tk) graphical library.
It requires the package `lbseed` that provides supporting code
for the interface.

It also uses the `lbrytools` library that implements
the `ch_download_latest` method to download the latest claims
from a channel.

https://https://github.com/belikor/lbrytools

If this function is added to the official `lbry-sdk` then we won't need
`lbrytools` anymore, only the main SDK (`lbrynet`).