lbrytools: add submodule for the `lbrytools` repository
The `lbrytools` repository is included as a subdirectory of this
project so that it is used in the graphical interface.

https://https://github.com/belikor/lbrytools

What is added as a submodule is only the `submodule` branch,
in which the code is placed in the toplevel directory:
```
git submodule add -b submodule https://github.com/belikor/lbrytools
```

Using the master branch as a submodule doesn't work smoothly,
as the `lbrytools` code is under two directories, that is,
`lbrytools/lbrytools/`.