lbseed: remove code that is used to create tooltips
This tooltip class will show a rectangle of text when the pointer
hovers on top of a specified widget.
```py
entry = tk.Entry()
tooltip.ToolTip(entry, text="popup text")
```

This code was used in prototype versions of this program
but it isn't used anymore so we remove it.

It is based on code included in IDLE, that is, `idlelib`.

https://hg.python.org/cpython/file/63a00d019bb2/Lib/idlelib/ToolTip.py