lbseed: separate various functions into a new module `resolve.py`
Modules that deal with resolving are moved from `res_download.py`
to a new module `resolve.py`.

The `res_download.py` module is also renamed to `actions.py`
as the methods which perform the final action in each page
are located here.

In the main program `dseed.py` we change the import appropriately
so that the code doesn't break:
```py
import lbseed.res_download as rs
```

becomes
```py
import lbseed.resolve as res
import lbseed.actions as actions
```