dseed: add a new page 2 to the `ttk.Notebook` to download single items
The first page is renamed from "Download" to "Download channel",
so the second page is called "Download single".

The old second and third pages are pushed to the third and fourth pages.

The widgets are renamed to have suffixes related to their topic,
for example, `_dch` for downloading channels, `_list` for
listing claims, `_del` for deleting claims.