dseed: encapsulate the creation of entry fields in a single function
The entry fields are generally uniform with a monospace font
therefore we can encapsulate the creation inside a function `setup_download_entry`
that is placed outside of the `Application` class.

The variables for the entry fields can be placed at a higher level in the class,
in `setup_widgets`.

Inside the pages we can set up the entry fields with a single function
```
    setup_download_entry(parent, server_var=self.server_var,
                         dir_var=self.down_dir_var,
                         font="Monospace 10", start=start)
```