dseed: encapsulate the creation of checkbuttons in a single function
The checkbuttons in the download pages are uniform with a monospace font
therefore we can encapsulate the creation inside a function `setup_download_check`
that is placed outside of the `Application` class.

The variables for the checkbuttons can be placed at a higher level in the class,
in `setup_widgets`.

Inside the pages we can set up the entry fields with a single function
```
    setup_download_check(parent, own_dir_var=self.own_dir_var,
                         save_var=self.save_var,
                         start=start+3)
```