dseed: add a new `page0` to include general configuration
A generic configuration option is the `lbrynet` listening server,
which is used by all functions.
By default it is `http://localhost:5279`.

The `setup_download_entry` function that is used in `page1` and `page2`
removes the server entry field, as it will appear only in `page0`.

The `start=` value is adjusted so it has one less row because
the entry field was removed.