dseed: encapsulate the creation of radiobuttons in a single function
The radiobuttons in the delete pages are uniform
therefore we can encapsulate the creation inside a function `setup_delete_radio`
that is placed outside of the `Application` class.

The variables for the radiobuttons can be placed at a higher level in the class,
in `setup_widgets`.

Inside the pages we can set up the radiobuttons with a single function
```
    setup_delete_radio(parent, del_what_var=self.del_what_var,
                       start=start)
```