dseed: encapsulate the creation of buttons in a single function
The buttons to validate input and resolve online are uniform
therefore we can encapsulate the creation inside a function `setup_grid_button_val_res`
that is placed outside of the `Application` class.

The validate and resolve functions are passed as arguments.

Inside the pages we can set up the buttons with a single function
```
    setup_grid_button_val_res(parent,
                              validate_func=self.validate_ch,
                              resolve_func=self.resolve_ch,
                              start=start)
```