dseed: print message after validation, and resolving
The validate button is used in order to make sure that the textbox
can be parsed correctly, either for downloading or for deleting
channel claims.

The button for resolving channels, and the button
for resolving claims are used to make sure that the channels
and claims exist in the network.

After the end of these operations an optional message `"Done"`
is printed to indicate that the operation is finished.