dseed: add more comments to the initialization of the application
The style is renamed from `them` to `theme`, and is moved before
initializing the appplication.

Clarify that the initialization of the application what actually does
is create a main frame in which all other widgets will be contained.
This is the reason the Application class is a derived class from
the `Frame` class.

Thus these two are the same
```
super().__init__(root)

mainframe = Frame(root)
```
and
```
self.pack()

mainframe.pack()
```
to show this main frame.

This main frame does not need to be packed at all, as long as
the children widgets use the `root` window as parent,
and are packed at some point after this.
```
note = ttk.Notebook(root)
note.pack()
```

However, we explictly use the `Frame` as parent, to indicate
more clearly the relationship.
```
note = ttk.Notebook(self)

note = ttk.Notebook(mainframe)
note.pack()
```