pages: move the mixing classes `Pages` to the submodule `pages.py`
The mixin classes (components) should not be used by themselves,
they are only used to extend the main `Application` class in `dseed.py`,
making the code shorter and easier to read.

The code could still be organized better because the methods associated
with the buttons in the mixin classes remain in the main `Application` class
which creates a twisted relationship in which the mixin classes
(components) aren't independent from the main class (composite).