pages.DownloadChPage: bind `"<<Activate>>"` to the button actions
The "Validate" and "Resolve channel" buttons are bound
to the "<<Activate>>" virtual event to launch
the `validate_func` and `resolve_func` functions.

The "Download claims" button is also bound to the virtual event
to launch the `download_ch` method.

All bound functions have to be wrapped with `blocks.f_with_event`
to make sure they accept an event, which is just discarded.

---- ---- ---- ---- ---- ---- ---- ----

For the entry field for the download directory
the "<<Activate>>" virtual event is bound to `focus_next_widget`
to move to the next widget.