pages.DownloadSinglePage: bind `"<<Activate>>"` to the button actions
The "Resolve claims" button is bound to the "<<Activate>>"
virtual event to launch the `resolve_func` function.

The "Download claims" button is also bound to the virtual event
to launch the `download_claims` method.

All bound functions have to be wrapped with `blocks.f_with_event`
to make sure they accept an event, which is just discarded.