pages.ListPage: bind `"<<Activate>>"` to the button actions
The "List claims" button is bound to the "<<Activate>>"
virtual event to launch the `list_claims` function.

The entry field that accepts a channel is also bound
to the virtual event to launch the same function.

All bound functions have to be wrapped with `blocks.f_with_event`
to make sure they accept an event, which is just discarded.