dseed: new `check_download_dir` to reset the download dir variable
This is placed in the `resolve.py` module.

The new method is used when pressing the `resolve_ch`
and `resolve_claims` buttons to download claims from channels
or individual claims.
If the download directory does not exist in the system,
it will use the default download directory from the configuration,
and will update the entry field before continuing.

The resolve functions are also called when doing the actual
download by `download_ch` and `download_claims`.