validate: eliminate the single quotation marks when we assume a channel
The `assume_channel` parameter will remove the single quotes
from a claim entry, that is, `'@claim-in-quotes'` to `@claim-in-quotes`.

However, this is not correct for video claims as many of them
include the single quote character as an apostrophe,
for example, `what's-cooking`, or `something-isn't-right`.
So, we will leave the quote intact.

Also, we will display the claim inside double quotes.