pages: add methods to enable and disable the checkboxes
In `DownloadChPage` and `DownloadSinglePage`
the second checkbox to place media files inside of a subdirectory
only makes sense when the first checkbox is enabled.

Therefore, we implement a method in which if the first checkbox
is checked, the second one can be checked or unchecked.
But if the first is unchecked, the second one is unchecked,
and disabled, so it cannot be checked.