dseed: enable and disable the checkbox when changing pages
When we move to another Notebook tab, it will run `update_checkbox`.
This method checks the tab we are on, and runs the appropriate
enable method for `DownloadChPage` or `DownloadSinglePage`.

The enable methods `chck_enable_dch` and `chck_enable_d`
have the option `force_second_var=False` so that the second
variable and checkbox is not changed to `True`
when changing Notebook tabs.

The second checkbox will be changed to `True` only when
the first checkbox is set to `True` and we are already
in the desired tab.