dseed: new options for printing claims in `ListPage`
Since the `lbrytools` library was updated, the `print_summary`
method now accepts three new parameters, `blocks`, `size`, and `sanitize`.

So two new checkboxes are used to control `blocks` and `size`.
The sanitization of the names is always `True`, as if the names,
titles or paths contain emojis, the Tk interface may crash.

The checkbox to show the `'claim_name'` is removed.
Instead, three radiobuttons are used to display either
the claim `'name'`, claim `'title'`, or claim `'path'`.

The `'path'` will only exist if the claim was completely downloaded
and the media file was created in the download directory.
Otherwise it will appear as `"None"`.