dseed: new checkbox for reversing the display of downloaded claims
To list the claims we use the new method `lbrytools.sort_items_size`
which returns the list of items but also the total size in bytes.

The new checkbox controls the state of the `reverse` parameter
in `lbrytools.print_items`.

A new label shows the total number of claims and the total size
in gibibytes (GiB) of the items.

This is added to two pages `ListPage` and `ListInvalidPage`.