dseed: add function to calculate size and duration
For `ListPage`, `ListInvalidPage`, and `ListChPage`
we calculate the size of all listed claims, and duration,
if they are of type video or audio; other types of claim
do not contribute to the duration value.

The `'duration'` is a new value that appears as the output
of the updated `lbrytools.sort_items_size`,
`lbrytools.ch_search_n_claims`, and `lbrytools.ch_search_all_claims`
methods.

The just like the `'size'` is displayed in GB for easier handling,
the `'duration'` is displayed in hours, minutes, seconds, and days.