pages_base: new label in the List pages to explain size and duration
This label is added to `ListPage`, `ListInvalidPage`, and `ListChPage`.

The 'size' corresponds to the size of the downloaded blobs;
if all media files were to exist the files would take double the space
on the hard drive.

The 'duration' considers only those claims that have a duration
such as video and audio files.