pages_base: add option to consider the claims invalid in `SupportAddPage`
A new variable `self.check_s_invalid` and checkbox
allows us to consider the claims as invalid.

There may be invalid claims that are still supported.
So with this checkbox we can change the underlying method
`lbrytools.abandon_support` to `lbrytools.abandon_support_inv`
to process these invalid claims, and change or completely abandon
the support on these claims.