dseed: new page `ListChsPeersPage` for listing peers of various channels
Get the peers from the latest claims of various channels
written in the upper textbox.
The channels are first resolved online, and if they exist
it procedes to find the latest claims from those channels.

Then it performs a peer search on each claim that can be downloaded,
that is, streams, as these are the only ones that can have peers.

Finally, it puts the summary of peers for each channel
in the lower textbox.

Searching for peers for a single claim takes considerable time,
so we can use threads to search many claims in parallel.

Since we want to search various channels, we can also use threads
to process channels in parallel.