dseed: new page `ListChPeersPage` for listing peers from a single channel
Get the peers from the latest claims of a given channel.
The channel is first resolved online, and if it exists
it procedes to find the latest claims from that channel.

Then it performs a peer search on each claim that can be donwloaded,
that is, streams, as these are the only ones that can have peers.

Finally, it puts the summary of peers and list of claims
with their peers in the textbox.

Searching for peers for a single claim takes considerable time,
so we can use threads to search many claims in parallel.