dseed: new button for `ListChSubsPage` to list the claims of subscribed channels
A spinbox specifies how many claims from each channel
will be listed.

The other radiobuttons for "shared" or "local" database,
and for showing "all", "valid", or "invalid" channels work the same.

The checkbox to show claim ID can be used to show the `'claim_id'`
of each individual claim under each channel.
The checkbox to show the title shows the `'title` of each claim,
as opposed to their claim `'name'`.

The spinbox to specify the number of threads works the same.

Also add a new image for the new page, and update the `README.md`.