ListChSubsPage: force `number=0` to `number=1`
If the `number` variable were to be 0, the internal function
`lbrytools.list_ch_subs_latest would search all claims
under a channel, which would take a long time.
Fortunately the internal function changes 0 to 1 by default
(if this is not overriden).

So we do the same change in the interface;
if the user inputs 0 in the spinbox, when calling the function
the spinbox will change the 0 to 1 to make it clear to the user
that 0 is not a permitted value.

In the interface there is no way to override this behavior.
In the internal `lbrytools.list_ch_subs_latest` method
this can be changed by setting `override=True`.