lbseed: update `validate.validate_input` to return a list of dictionaries
The `validate.validate_input` is used to parse the text
in a textbox to get a pair consisting of a channel name
and a number.
```
@chan1, 43
@channel2, 16
@channel4, 3
```

Previously the information returned would be two lists,
`channels` and `numbers`. Now `validate.validate_input` is changed
so that we return a single list, with each element being a dictionary
with two keys `'claim'` and `'number'`.

This output list `validated_chs` is in turn passed to `resolve.resolve_ch`
to verify that the channels exist.

Previously `resolve.resolve_ch` returned only one list `resolve_info`
with the information from the resolution of each channel.
Now we return an augmented list of dictionaries `resolved_chs`,
where each dictionary has three keys, `'claim'`, `'number'`,
and `'info'`.

The output list with validated and resolved channels `resolved_chs`
can then be passed to the functions that do operations
with the channels, including `act_download.download_ch`
and `act_delete.clean_ch`.

These two methods are modified to extract the proper channel name
and number from the dictionaries, as well as the information
that tells us if the channel is valid or not.

The `validate.validate_input` and `resolve.resolve_ch` methods
are called from `validate_ch`, `resolve_ch`, and
these are used by `download_ch` and `delete_ch`.

----

The new `resolve.resolve_ch` method is also used in `resolve_ch_list`.

In this case it's only one channel, so we build a small list
with only a single element, with the channel name,
and a number equal to `False` (the number isn't used anyway).

The returned `resolved_chs` is a list with a single element
which has the three keys, `'claim'`, `'number'`, and `'info'`.
The latter indicates if the channel is actually valid or not.

If the channel is valid `resolve_ch_list` returns the channel name
from `'info'`, which is then passed to `act_list.print_ch_claims`
inside `list_ch_claims`.

----

The new `validate.validate_input` is used to parse pairs claims
and numbers to produce `validated_claims`.
This `validated_claims` is used inside `resolve_g_claims`
which is now passed to `resolve.resolve_claims_supports`.
This method takes each dictionary in the list and expands it
so that it has a third key `'resolved'` with the resolved
item indicated by the `'claim'` key.

The `resolved_claims` list is in turned passed to `act_supports.add_supports`
where the keys `'resolved'` and `'number'` are used
to add, create, abandon, change, or target a specific support.