dseed: new page `ListSubsPeersPage` for listing peers of subscribed channels
Get the peers from the latest claims of all channels
that are in our subscription list.
The channels are resolved online, and if they exist
it procedes to find the latest claims from all those channels.

Then it performs a peer search on each claim that can be downloaded,
that is, streams, as these are the only ones that can have peers.

Finally, it puts the summary of peers for each channel
in the textbox.

Searching for peers for a single claim takes considerable time,
so we can use threads to search many claims in parallel.

Since we search various channels, hundreds even, we can also use threads
to process channels in parallel.