README: move information to documents in `doc/` subdirectory
The `README.md` now has a list of content, going to the individual
sections in the documentation, 'Introduction', 'Installation',
'Usage', and 'Development'.

The majority of the 'Installation' is placed in its own document
`doc/installation.md`.

The majority of the 'Usage' is placed in its own document
`doc/usage.md`.

Also add a .desktop file, and .svg and .png icons
so that they can be used in some graphical system,
for example, an AppImage.