dseed: rename class `ListPage` to `ListDownPage`
And `ListInvalidPage` to `ListDownInvalidPage`,
and the associated commands to build the interface.

We do this to be consistent, as these pages refer to claims
that have already been downloaded.

----

Also rename method from `list_claims` to `list_d_claims`,
and `list_inv_claims` to `list_d_claims_inv`.

The methods refers specifically to downloaded claims so we add `d`
to the name to stress this.

In another commit we will list generic claims that we have published
so we need to distinguish different types of claims.

----

Move the `update_checkbox` closer to the downloading methods
before `validate_ch`, `resolve_ch`, `download_ch`, `resolve_claims`,
and `download_claims`.

----

Do small adjustments to the `act_download` methods.