dseed: new page `ListPubChsPage` to display our created channels
These channels were created with our default wallet.

We print a sumary of all channels created, and the amount of LBC
staked in the channels.

Various fields of information can be displayed per channel including
time of last update, claim ID, addresses of the transaction,
account of the transaction, and LBC staked.