dseed: new page `ListPubClaimsPage` to display our published claims
These claims were created with our default wallet.

We print a summary of all claims published, the number of channels,
total claims in the channels, total size, and total duration
of the claims.

It will also display a summary of claims published anonymously,
that is, not inside any channel.

Various fields of information can be displayed per claims including
time of last update, claim ID, addresses of the transaction,
type of claim, stream and media, if available, LBC staked,
and name or title of the claim.