dseed: rename class `ListChSubsPage` to `SubscribedChsPage`
The page displays the subscribed channels, and the claims
inside the subscribed channels. The longer name makes it clearer
what it is.

----

We also rename the page `page_ch_subs` to `page_subscr_chs`.

We also rename the methods `list_ch_subs` to `list_subscr_chs`
and `list_ch_subs_claims` to `list_subscr_chs_claims`.