pages: separate support variables into a mixin class
The variables for the support pages are placed
in the `VarsSupports` class which is used as a mixin class
for the `Variables` class.

----

The methods used in the `SupportListPage` and `SupportAddPage`
are moved together.

This means `validate_g_claims`, `resolve_g_claims`,
and `add_supports` now come immediately after `list_supports`.