pages: separate trending variables into a mixin class
The variables for the trending and search pages are placed
in the `VarsTrending` class which is used as a mixin class
for the `Variables` class.

---

Also the variable `chck_tr_claim_t` is renamed to `rad_tr_claim`
because the value is used in a group of radiobuttons.

---

The methods used in the `TrendPage` and `SearchPage`
are moved together.

This means `update_search_checkbox`, `show_trending_claims`,
and `show_search` now come immediately after `add_supports`.