pages: separate controlling claims variables into mixin class
The variables for the controlling claims page are placed
in the `VarsControlling` class which is used as a mixin class
for the `Variables` class.

---

Also rename `act_others.show_claims_bids` to `act_others.ctrl_claims`
so that the method is shorter.

Also shorten the name of the parameters `show_controlling` to `show_contr`,
`show_non_controlling` to `show_non_contr`,
and `show_repost_status` to `show_repost_st`.

This is done so that the lines aren't very long.