blocks: rename `setup_spin_page` to `setup_spin_gen`
This method sets up a generic spinbox widget for integer numbers
which means it can be used to specify quantities such as a page,
a number of claims, a number of threads, and others.

The generic method is moved to `blocks_base.py`. Now the method
also accepts the values of from, to, the increment, and the default
value of the spinbox.

The new command is used in existing pages
- `ListChClaimsPage` to select the number of claims

- `SubscribedChsPage` to select the number of claims
- `SubscribedChsPage` to select the number of threads for channels

- `ListChPeersPage` to select the number of claims
- `ListChPeersPage` to select the number of threads for claims

- `ListChsPeersPage` to select the number of threads for channels
- `ListChsPeersPage` to select the number of threads for claims

- `ListSubsPeersPage` to select the number of claims
- `ListSubsPeersPage` to select the number of threads for channels
- `ListSubsPeersPage` to select the number of threads for claims