dseed: new page `CommentsPage` to display comments from a claim
The input is a claim, that could be a claim name or claim ID.
It can also be a channel name, `@channel`.

It prints the newest 600 comments below that claim in a listbox,
including the threaded replies under each comment.

Pressing a button to reply will open the editor
to create a comment reply, edit a previous comment,
or abandon a previous comment.