dseed: move the page "Seeding ratio" to the main "Peers" tab
The "Seeding ratio" page is better being alongside other pages
that deal with peers, "Channel peers", "Multiple channel peers",
and "Subscription peers", so it is placed in the main "Peers" tab.

----

The corresponding class `SeedPage` is moved from `pages_adv.py`
to `pages_peers.py`, and the path is adjusted in `pages.py`.

The corresponding method `seeding_ratio` is moved from `act_others.py`
to `act_peers.py`, and the path is adjusted in `actions.py`.