dseed: move the page "Controlling claims" to the main "List claims" tab
The "List claims" tab has other pages that list different claims,
so it fits better there.

Some pages where also renamed:
- "List downloaded claims" to "Downloaded claims"
- "List invalid claims" to "Invalid claims"
- "List channel claims" to "Channel claims"
- "Subscribed channels"
- "Created channels"
- "Published claims"
- "Controlling claims"

The "Controlling claims" page checks whether the claims we have published
are the "controlling" ones, meaning that have the highest bids
for their given "claim name".

If they don't have the highest bids, they won't be found at the top
of the search results. So if we want them to appear first,
we need to add a bid that surpasses the highest bid on the list.

----

The corresponding class `ControllingClaimsPage` is moved from `pages_adv.py`
to `pages_lists.py`, and the path is adjusted in `pages.py`.

The corresponding method `ctrl_claims` is moved from `act_others.py`
to `act_list.py`, and the path is adjusted in `actions.py`.

The `act_others.py` module is kept empty. It could be used
in the future to put methods that don't fit in other modules.

----

The existing `page_adv.py` is removed.
A new module is added `pages_other.py` but this remains empty.
It could be used in the future for other page classes
that do not fit in the existing categories.

Likewise the "Advanced" tab is renamed to "Other" in the interface
but is not actually displayed. It is just a space that could be used
in the future to add interfaces that do not fit well in other tabs.