doc: updated images for the usage documentation
New image for the new comments page:

- `g_lbrydseed_comments.png`

Renamed images:

- `g_lbrydseed_controlling.png` to `g_lbrydseed_controlling_claims.png`
- `g_lbrydseed_seeding.png` to `g_lbrydseed_seeding_ratio.png`
- `g_lbrydseed_list_supports.png` to `g_lbrydseed_supports_list.png`
- `g_lbrydseed_add_support.png` to `g_lbrydseed_supports_add.png`

Updated images:

- `lbrydseed-ch-claims.png`
- `g_lbrydseed_download_channel.png`
- `g_lbrydseed_download_single.png`
- `g_lbrydseed_list_down_claims.png`
- `g_lbrydseed_list_down_invalid.png`
- `g_lbrydseed_list_ch_claims.png`
- `g_lbrydseed_subscribed_channels.png`
- `g_lbrydseed_created_channels.png`
- `g_lbrydseed_published_claims.png`
- `g_lbrydseed_peers_channel.png`
- `g_lbrydseed_peers_multi_channels.png`
- `g_lbrydseed_peers_subscriptions.png`
- `g_lbrydseed_delete_single.png`
- `g_lbrydseed_cleanup_channel.png`
- `g_lbrydseed_trending.png`
- `g_lbrydseed_search.png`

In the program the `Advanced` page was removed,
its two subpages `Controlling claims` and `Seeding ratio`
were moved to other pages.
The first, `Controlling claims`, is moved at the end of `List claims`,
and the second, `Seeding ratio`, is moved at the end of `Peers`.

The images in `usage.md` are moved from the `## Advanced` section
to `## Listing` and `## Peers`, and the `## Advanced` section
is removed.