dseed: rename `print_claims` to `list_claims` and refactor
The original method to list downloaded claims
and downloaded invalid claims is `act_list.print_claims`;
this is renamed to `act_list.list_claims`.

The original method to list claims from a channel
is `act_list.print_ch_claims`;
this is renamed to `act_list.list_ch_claims`.

Originally the `act_list.list_text_size` method is exposed
in `actions.py`, and called in `dseed.py`.
This is no longer needed as it is only used internally
by `act_list.list_claims` and `act_list.list_ch_claims`.

The `list_text_size` method instead of accepting three parameters
`number`, `size`, `seconds`, it accepts a single dictionary
from which this information is extracted.

The labels that have information of claims, the size,
and duration are removed from `pages_lists.py`.
The corresponding variables that hold this information
`label_lst_info`, `label_lst_inv_info`, and `label_chl_info`
are removed from `variables.py`.

This information is now calculated in the `list_claims`
and `list_ch_claims` methods; it is concatenated
with the list of claims, and printed inside the textbox.