create_appimage: script to automatically create an `AppImage`
The script will create a directory where it will place the tools
necessary to assemble the AppImage, `appimagetool`, `linuxdeploy`,
`linuxdeploy-plugin-python`, and `appimage-builder` (not used).

The `linuxdeploy` program will be used with the Python plugin
to create a base `AppDir` directory that will contain a basic
installation of Python, including the required packages
such as `requests`, `emoji`, `regex`, `Pillow` and `matplotlib`.

Inside this directory the submodules `lbrytools` and `lbseed`,
and the main executable `dseed.py` will be placed.
The AppImage also requires some additional files
such as a `.desktop` file, and the icon `.svg`.
A `lbrydseed.png` image is placed next to `dseed.py`
as it will be used by it to add the icon to the title bar
of the interface.

A custom `AppRun` script will be created in `AppDir`
in order to call `dseed.py` with the internal Python executable.
The `AppRun` script is the entry point when calling the AppImage.

Finally, `appimagetool` is called on the entire `AppDir`
to produce the final AppImage.