dseed: if `content` is False, indicate that no claims were found
If `actions.list_ch_claims` (which calls `lbrytools.ch_search_n_claims` or
`lbrytools.ch_search_all_claims`) doesn't return claims for whatever reason
(including bad Internet connection to the hub servers), the output dictionary
will have a `'content'` value that is `False`.

In this case, set it to `'No claims found'`, which will be displayed
in the textbox.