SupportAddPage: fix `add_supports` when choosing invalid claims
When we enable the invalid checkbox, it will assume the claim,
video or channel, is not valid any more, so it will not resolve it
online. This must be done to change or abandon the support
of invalid claims.

When the return value of `validate.validate_input`
and `resolve.resolve_claims_supports` were changed
in c2fb514c, the output dictionary was not properly handled
in the case that the invalid checkbox is checked.

If the invalid checkbox is checked, `resolve_g_claims`
will obtain a `validated_claims` list from `validate_g_claims`
but it won't return a `resolved_claims` list,
so `resolve_g_claims` will fail.

So for this case, the `resolved_claims` is simply set
to `validated_claims`.

----

To perform the action `resolved_claims` is passed
to `act_supports.add_supports`.

When the invalid checkbox is checked, the `claim` variable
will not be defined because it is not used by other methods
to get the name or claim ID.

In that case, the `name` will simply be set to the value
of the `'claim'` key, which may be a claim name or a claim ID,
depending on what the user wrote in the textbox.

To complete the operation, the method `lbrytools.abandon_support_inv`
will be called with the appropriate input, first with `cid=name`,
and then with `name=name`, if the first call fails.