pages_support: add spinbox to choose the number of threads
When listing supports, each claim is resolved online
to get the full information on the supports.
Threads make the process faster as multiple claims are resolved
in parallel.

Add a new variable `self.spin_s_threads` and we pass it
to the two main methods `list_supports` and `add_supports`,
which underneath call `act_supports.list_supports`
and `act_supports.add_supports`, which call corresponding methods
in `lbrytools`.