act_list: use the new output of `lbrytools.print_summary`
With the new updated `lbrytools` now we can take the output
directly from `lbrytools.print_summary`, instead of using
`lbrytools.sort_items_size` and then `lbrytools.print_items`,
and manually calculating the size and duration of the claims.

We also add the new `threads` parameter, which will speed up
resolving of the claims to display the invalid claims.