pages_lists: add spinbox to choose the number of threads
When listing invalid claims, first all downloaded claims are resolved online
to get the full information on the claims. If the resolution fails
it means the failed claim is in fact invalid.
Threads make the process faster as multiple claims are resolved
in parallel.

Add a new variable `self.spin_lst_threads` and we pass it
to the main methods `list_d_claims` and `list_d_claims_inv`,
which underneath call `act_list.list_claims`,
which calls the `lbrytools.print_summary` method.