act_peers: use new methods for peer search
The `lbrytools` library was updated to include new methods
and parameters.

Some of the existing methods were renamed:
- `list_peers`: it is a new method to search peers of a single claim.
- `list_ch_peers`: the existing `list_peers` was renamed to `list_ch_peers`
  as it searches the claims of a single channel.
- `list_chs_peers`: the existing `list_ch_peers` was renamed to `list_chs_peers`
  as it searches the claims of multiple channels.

In the past we would construct the `summary` variable manually.
Now this summary is already provided as the output of the corresponding
method `list_ch_peers`, `list_chs_peers`, and `list_ch_subs_peers`.

The summary is located inside the `'summary'` key of the output dictionary.