pages_peers: new variable to display peer information as paragraph
With a new checkbox we can choose to display the peer information
of each claim as a full paragraph, or just as a single line
as originally.

The new checkbox when turned on deactivates the other three checkboxes
for the claim ID, type, and title, as when the paragraph is displayed
these values are always displayed as well.