pages_peers: new page to find the peers of arbitrary claims
The page `ListChPeersPage` allows us to search peers from the claims
of a single channel, the page `ListChsPeersPage` allows us to search peers
from the claims of multiple channels, and the page `ListSubsPeersPage`
allows us to search peers from the claims of our subscribed channels.

A new page `ListClsPeersPage` is added to search peers from a list
of arbitrary claims, either by URI or by claim ID.

The page uses the method `lbrytools.list_m_peers` with a properly
resolved list of claims, meaning that only those claims that
exist will be actually searched.

The same checkboxes that exist for `ListChPeersPage` are used
in `ListClsPeersPage`, to display claim ID, media type, title,
and optionally a paragraph of information for each claim.

A small method `update_peers_checkbox` is called each time the notebook
changes sub tabs from `ListClsPeersPage` ("Claim peers")
to `ListChPeersPage` ("Channel peers") in order to update
the states of the checkboxes, as if the last checkbox is active
the other checkboxes do nothing.