pages_search: add new checkboxes for release time and title
With the new methods `lbrytools.list_trending_claims`
and `lbrytools.list_search_claims`, now they accept a `release`
and `title` parameters to display the release time (or creation time)
of the claim, or the title of the claim instead of the name.

----

The `TrendPage` and `SearchPage` classes in `page_search.py`
are also updated with new checkboxes to control the display
of release time (or creation time), and title.

New variables are introduced in the class `VarsSupports` in `variables.py`
for these checkboxes.