pages_search: set up the `page=0` by default to get 1000 claims maximum
With the new methods `lbrytools.list_trending_claims`
and `lbrytools.list_search_claims` now the `page` parameter
can also be zero or `None`, in which case the methods will display
all 20 pages, for a maximum of 1000 claims.

----

In the `TrendPage` and `SearchPage` classes in `page_search.py`
that hold the interface, the page spinbox now goes from 0 to 20,
and defaults to 0.

In `variables.py` the `spin_page` variable is set to 0 by default
as well.