pages_search: add new spinbox to select number of threads
The new methods `lbrytools.list_trending_claims`
and `lbrytools.list_search_claims` now accept a `threads` parameter
to define the number of threads with which to process the 20 pages
of the search.

----

In the `TrendPage` and `SearchPage` classes in `page_search.py`
a new spinbox controls the number of threads, which defaults to 32.

In `variables.py` the `spin_tr_threads` variable is set to 32 by default.