pages_search: add more information to the list of searched claims
With the new methods `lbrytools.list_trending_claims`
and `lbrytools.list_search_claims`, now the return value
is a dictionary with more information, including the input parameters
of the search, and also a summary of the claims found.

This information is returned from the corresponding methods
in `act_search.py`, and they are added to the printed lines
in the main textbox.

----

In the interface classes `TrendPage` and `SearchPage` in `pages_search.py`
the label for the searched page is removed, as this information
is provided now in the textbox.

The corresponding label variables in `VarsTrending` in `variables.py`
are also removed since they are not needed anymore.

----

In the interface classes `TrendPage` and `SearchPage`
the paragraph that explains the default returned information is augmented.
This paragraph is typed in the `blocks_trending.py` module.

The new methods `lbrytools.list_trending_claims`
and `lbrytools.list_search_claims` now also indicate the fee
that needs to be paid for accessing the claims, if any.