act_list: use new output value for `lbrytools.print_summary`
The `'text'` key in the return dictionary of `lbrytools.print_summary`
was renamed to `'summary'`, as it's a summary of the claims processed,
including size and duration.

Therefore the corresponding methods in `act_list.list_claims`
needs to be updated as well to use the `'summary'` key,
and not the `'text'` key.

Also rename the dictionary key `'content'` to `'lines'`
to be more consistent with other methods, as this key will have
the lines to be listed in the textbox.